requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain and AI (Latest versions)
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.30
openai>=1.30.0

# Vector Database - Updated to newer package
langchain-chroma>=0.1.0
sentence-transformers>=2.2.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0

# Streaming responses
sse-starlette>=1.6.0

# Rate limiting
slowapi>=0.1.9

# Web search capabilities
aiohttp>=3.8.0

# Dev tools
black>=24.3.0
ruff>=0.5.0
mypy>=1.10.0
isort>=5.13.2